Move compat tests to a folder
