+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Apache.Ignite.Core.Tests.csproj
@@ -168,10 +168,8 @@
     <Compile Include="Client\ClientProtocolCompatibilityTest.cs" />
     <Compile Include="Client\ClientFeaturesTest.cs" />
     <Compile Include="Client\ClientProtocolVersionTest.cs" />
-    <Compile Include="Client\ClientReconnectCompatibilityTest.cs" />
     <Compile Include="Client\ClientServerCacheAdapter.cs" />
     <Compile Include="Client\ClientServerCacheAdapterExtensions.cs" />
-    <Compile Include="Client\ClientServerCompatibilityTest.cs" />
     <Compile Include="Client\ClientTestBase.cs" />
     <Compile Include="Client\Cluster\ClientClusterDiscoveryTests.cs" />
     <Compile Include="Client\Cluster\ClientClusterDiscoveryTestsBase.cs" />
@@ -180,6 +178,8 @@
     <Compile Include="Client\Cluster\ClientClusterDiscoveryTestsSsl.cs" />
     <Compile Include="Client\Cluster\ClientClusterGroupTests.cs" />
     <Compile Include="Client\Cluster\ClientClusterTests.cs" />
+    <Compile Include="Client\Compatibility\ClientReconnectCompatibilityTest.cs" />
+    <Compile Include="Client\Compatibility\ClientServerCompatibilityTest.cs" />
     <Compile Include="Client\Compute\ComputeClientDisabledTests.cs" />
     <Compile Include="Client\Compute\ComputeClientTests.cs" />
     <Compile Include="Client\EndpointTest.cs" />
