+++ modules/platforms/dotnet/Apache.Ignite.Core.Tests/Client/Compatibility/ClientReconnectCompatibilityTest.cs
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-namespace Apache.Ignite.Core.Tests.Client
+namespace Apache.Ignite.Core.Tests.Client.Compatibility
 {
     using System;
     using System.Linq;
@@ -28,7 +28,7 @@ namespace Apache.Ignite.Core.Tests.Client
     /// <summary>
     /// Tests thin client compatibility with reconnect.
     /// </summary>
-    [Category(TestUtils.CategoryIntensive)]    
+    [Category(TestUtils.CategoryIntensive)]
     public class ClientReconnectCompatibilityTest
     {
         /// <summary>
@@ -43,7 +43,7 @@ namespace Apache.Ignite.Core.Tests.Client
                 EnablePartitionAwareness = true,
                 Logger = new ListLogger(new ConsoleLogger {MinLevel = LogLevel.Trace})
             };
-            
+
             try
             {
                 using (StartNewServer())
